EES-4771 General cleanup before resuming the great migration to Azure.Data.Tables
